TECH-336: Move shared types back into component folders
