+++ src/components/Button/Button.spec.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {shallow} from 'component-test-utils-react';
 import {Button} from './index';
-import {buttonColors, buttonSizes, buttonVariants} from '~/shared-types/button.types';
+import {buttonColors, buttonSizes, buttonVariants} from './Button.types.ts';
 import Love from '~/icons/Love';
 
 describe('Button', () => {
