+++ src/components/PrimaryNav/PrimaryNavItem/PrimaryNavItem.jsx
@@ -2,8 +2,7 @@ import React, {useContext} from 'react';
 import PropTypes from 'prop-types';
 import './PrimaryNavItem.scss';
 import classnames from 'clsx';
-import {Typography} from '~/components/Typography';
-import {variants as typographyVariants} from '~/shared-types/typography.types';
+import {Typography, variants as typographyVariants} from '~/components/Typography';
 import {PrimaryNavContext} from '../PrimaryNav.context';
 
 // Internal component
