+++ src/components/Typography/Typography.stories.jsx
@@ -4,8 +4,7 @@ import {storiesOf} from '@storybook/react';
 import {boolean, select, withKnobs} from '@storybook/addon-knobs';
 import storyStyles from '~/__storybook__/storybook.module.scss';
 
-import {Typography} from './index';
-import {variants, weights} from '~/shared-types/typography.types';
+import {Typography, variants, weights} from './index';
 import markdownNotes from './Typography.md';
 
 storiesOf('Tokens/Typography', module)
