+++ src/components/Typography/Typography.tsx
@@ -1,7 +1,12 @@
 import React, {FunctionComponent} from 'react';
 import classnames from 'clsx';
 import './Typography.scss';
-import {TypographyWeight, TypographyVariant} from '~/shared-types/typography.types';
+
+export type TypographyVariant = 'title' | 'heading' | 'subheading' | 'body' | 'caption' | 'button';
+export const variants = ['title', 'heading', 'subheading', 'body', 'caption', 'button'];
+
+export type TypographyWeight = 'default' | 'bold' | 'semiBold' | 'light';
+export const weights = ['default', 'bold', 'semiBold', 'light'];
 
 interface TypographyProps {
     /**
