--- src/shared-types/button.types.ts
@@ -1,8 +1,0 @@
-export type ButtonSize = 'small' | 'default' | 'big';
-export const buttonSizes = ['small', 'default', 'big'];
-
-export type ButtonVariant = 'default' | 'ghost' | 'outlined';
-export const buttonVariants = ['default', 'ghost', 'outlined'];
-
-export type ButtonColor = 'default' | 'accent' | 'danger';
-export const buttonColors = ['default', 'accent', 'danger'];
