--- src/shared-types/typography.types.ts
@@ -1,5 +1,0 @@
-export type TypographyVariant = 'title' | 'heading' | 'subheading' | 'body' | 'caption' | 'button';
-export const variants = ['title', 'heading', 'subheading', 'body', 'caption', 'button'];
-
-export type TypographyWeight = 'default' | 'bold' | 'semiBold' | 'light';
-export const weights = ['default', 'bold', 'semiBold', 'light'];
